Write credentials to /home/core/.dockercfg.
The only credentials file mounted within the mesos slave container is
/home/core/.dockercfg.  The previous method failed because newer docker
versions write credentials to ~/.docker/config.json instead of
~/.dockercfg.

This forced me to find a more robust approach to the creation of
.dockercfg since aws cli would not work as it only spits out the 'docker
login'.  Fortunately I found a solution in a github project and
dockerhub image, which I forked and added onto.

Details can be found here https://github.com/adobe-platform/ecr-login.
The tool fetches the credentials and places them in the .dockercfg
format.  The aws-ecr-login service overwrites /home/core/.dockercfg with
the output.

The .dockercfg for the behance dockerhub registry remain in
/root/.dockercfg and are used by fleetctl/systemd services even when
other users are specified in the service files.  Therefore this does not
interfere with other services that my need those credentials.  It will
interfere with any containers run through Marathon that are not found in
the chosen ECR registry.  If desired, this challenge could be
circumvented by allowing the specification of the Marathon URI field in
FlightDirector.  The URI could be directed at the URL of an S3 bucket
and secured with the S3 container policy.  This could be automated with
an additional fleetctl service which uses the ecr-login container
with the added modification of s3 push.

The alternative is to have all images exist in a single ECR registry.
IMHO the best way to manage this is for PDF Services build automation to
push to that ECR registry.  It's very simple to adjust the permissions on
a per repository basis in ECR.

That discussion aside, when we cease using DockerHub the ecr-login
Docker image will need to pulled using alternate means.  The simplest
way I can think of is to store it as a tarball in an S3 bucket which is
pulled by the worker instances at boot.